requirements.txt
python-dotenv
google-generativeai
streamlit
streamlit-authenticator
streamlit-lottie
langchain-google-genai
ipykernel
langchain-community
langchain
PyPDF2
typing-extensions
pypdf
faiss-cpu